# src/main/resources/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Lead Service API
  version: 1.0.0
  description: API для управления лидами (потенциальными клиентами)

tags:
  - name: Lead

servers:
  - url: http://localhost:8081
    description: Local development server

paths:
  /api/v1/leads:
    post:
      tags:
        - Lead
      summary: Create a new lead
      operationId: createLead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeadCreateRequest'
      responses:
        '201':
          description: Lead created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadResponse'
        '400':
          description: Validation error (missing title or personId)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - Lead
      summary: List leads by status
      operationId: listLeads
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/LeadStatus'
      responses:
        '200':
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeadResponse'
  /api/v1/leads/{id}:
    get:
      tags:
        - Lead
      summary: Get lead by ID
      operationId: getLeadById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lead found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadResponse'
        '404':
          description: Lead not found
    put:
      tags:
        - Lead
      summary: Update lead
      operationId: updateLead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeadUpdateRequest'
      responses:
        '200':
          description: Lead updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadResponse'
        '400':
          description: Validation error (empty title)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Lead not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Lead
      summary: Delete lead
      operationId: deleteLead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Lead deleted
        '404':
          description: Lead not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/leads/{id}/status:
    patch:
      tags:
        - Lead
      summary: Change lead status
      operationId: changeStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  $ref: '#/components/schemas/LeadStatus'
      responses:
        '200':
          description: Status changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadResponse'
        '400':
          description: Invalid status value
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Lead not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Invalid status transition
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    LeadStatus:
      type: string
      enum:
        - NEW
        - CONTACTED
        - QUALIFIED
        - DISQUALIFIED
    LeadCreateRequest:
      type: object
      required:
        - title
        - personId
      properties:
        title:
          type: string
        description:
          type: string
        personId:
          type: string
          format: uuid
        source:
          type: string
    LeadUpdateRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
    LeadResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          type: string
        personId:
          type: string
          format: uuid
        source:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        service:
          type: string
        errorCode:
          type: string
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string